--- matjalalzz/src/main/java/shop/matjalalzz/domain/user/domain/enums/Role.java
@@ -1,6 +1,0 @@
-package shop.matjalalzz.domain.user.domain.enums;
-
-public enum Role {
-    ADMIN, USER, OWNER
-
-}
