+++ matjalalzz/src/main/java/shop/matjalalzz/domain/user/entity/enums/Role.java
@@ -1,0 +1,6 @@
+package shop.matjalalzz.domain.user.entity.enums;
+
+public enum Role {
+    ADMIN, USER, OWNER
+
+}
