+++ matjalalzz/src/main/java/shop/matjalalzz/global/security/jwt/dto/TokenBodyDto.java
@@ -1,7 +1,7 @@
 package shop.matjalalzz.global.security.jwt.dto;
 
 import lombok.Builder;
-import shop.matjalalzz.user.domain.enums.Role;
+import shop.matjalalzz.domain.user.entity.enums.Role;
 
 @Builder
 public record TokenBodyDto(Long userId, String email, Role role) {
