+++ matjalalzz/src/main/java/shop/matjalalzz/global/security/jwt/utils/TokenMapper.java
@@ -3,8 +3,8 @@ package shop.matjalalzz.global.security.jwt.utils;
 import shop.matjalalzz.global.security.jwt.dto.LoginTokenResponseDto;
 import shop.matjalalzz.global.security.jwt.dto.TokenBodyDto;
 import shop.matjalalzz.global.security.jwt.entity.RefreshToken;
-import shop.matjalalzz.user.domain.User;
-import shop.matjalalzz.user.domain.enums.Role;
+import shop.matjalalzz.domain.user.entity.User;
+import shop.matjalalzz.domain.user.entity.enums.Role;
 
 public class TokenMapper {
     public static RefreshToken toRefreshToken(String refreshToken, User user) {
